optimised code and created an ADDSALE button in Sales list

+++ Frontend/src/app/sales/components/sales-list/sales-list.component.ts
@@ -11,30 +11,34 @@ import { Router } from '@angular/router';
   styleUrls: ['./sales-list.component.scss']
 })
 export class SalesListComponent implements OnInit {
-  @ViewChild(MatPaginator) paginator!: MatPaginator;
-  constructor(public sharedService: SharedService, public saleService: SalesService, private router: Router) { }
   salesList!: MatTableDataSource<any>;
-  displayedColumns = ["sell_id", "vehicle_no", "fullName", "sold_date", "sold_amount", "balance_amount", "actions"]
+  displayedColumns = ["sNo","vehicle_no", "fullName", "sold_date", "sold_amount", "balance_amount", "actions"];
+  constructor(public sharedService: SharedService, public saleService: SalesService, public router: Router) { }
   ngOnInit(): void {
     this.sharedService.getSalesList().subscribe((res: any) => {
-      if(res.data.length){
         this.sharedService.salesData = res.data;
-        this.salesList = new MatTableDataSource(this.sharedService.salesData);
-      }
+    this.salesList = new MatTableDataSource(this.sharedService.salesData);
     }, err => this.sharedService.snackbarNotification(err.error.message, "OK", {
       duration: 3000,
       panelClass: ['snackbar-fail']
     }));
   }
-  ngAfterViewInit() {
-    if(this.sharedService.salesData.length)
-    this.salesList.paginator = this.paginator;
+  @ViewChild(MatPaginator) set matPaginator(paginator: MatPaginator) {
+    if(paginator != undefined)
+    {
+    this.salesList.paginator =paginator;
+
+    }
   }
   editSale(index: any) {
     this.saleService.index = index;
     this.router.navigateByUrl("/admin/sales/editsale");
   }
   deleteSale(sell_id: any) {
+    if(!confirm("Are you sure to delete this sale"))
+    {
+      return;
+    }
     this.saleService.deleteSale(sell_id).subscribe((res: any) => {
       this.sharedService.snackbarNotification(res.message, "OK", {
         duration: 3000,
